+++ cpm/hierarchical/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["EmpiricalBayes"]
+
+from .empirical import EmpiricalBayes